% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{analyzeFeature}
\alias{analyzeFeature}
\title{Analyze a given feature}
\usage{
analyzeFeature(
  data = NULL,
  columnsToUse = NULL,
  feature = NULL,
  metricToUse = "median"
)
}
\arguments{
\item{data}{A data frame containing the dataset to export. This argument should be the output of the \code{bindData()} function. Defaults to \code{NULL}.}

\item{columnsToUse}{A character vector defining the columns considered as the variable of interest (typically cell clusters). Defaults to \code{NULL}.}

\item{feature}{A character vector defining the column to be used as a feature (typically group, cluster or metadata). Defaults to \code{NULL}.}

\item{metricToUse}{A string defining the metric to use for statistics computation. It can be either \code{mean} or \code{median}. Defaults to \code{median}.}
}
\value{
Return value is directly passed to \code{exportFeaturesTables()} function.
}
\description{
This function allows to analyze a given feature. As this function is called by \code{exportFeaturesTables()} function, it should not be called as is.
}
