% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Normalization.R
\name{analyzePeaks}
\alias{analyzePeaks}
\title{Analyze the peaks for each sample and parameter}
\usage{
analyzePeaks(
  parametersToAnalyze = NULL,
  max.lms.sequence = NULL,
  suffix = NULL,
  samplesToDelete = NULL,
  nCoresToExploit = NULL,
  minpeakdistance = 150,
  minpeakheight = 0.1
)
}
\arguments{
\item{parametersToAnalyze}{A vector of all parameters to analyze. Defaults to \code{NULL}.}

\item{max.lms.sequence}{A named numeric list containing the maximum expected/desired number of peaks for each parameter. Defaults to \code{NULL}.}

\item{suffix}{A string defining the pattern of which \code{rds} files to use. Defaults to \verb{_raw}.}

\item{samplesToDelete}{A character vector defining the samples to remove from the peaks analysis. Defaults to \code{NULL}.}

\item{nCoresToExploit}{An integer defining the number of cores to use to analyze the peaks. It corresponds to the number of parameters to be treated concomitantly. Defaults to \code{NULL}, which translates to the actual number of cores of the processor minus 1.}

\item{minpeakdistance}{An integer defining the minimum possible distance (in indices) between two peaks. The user should be careful if changing this value (see \code{findpeaks()} method from \code{pracma} package for more information). Defaults to \code{150}.}

\item{minpeakheight}{An integer defining the minimum possible absolute height for a peak. The user should be careful when changing this value (see \code{findpeaks()} method from \code{pracma} package for more information). Defaults to \code{0.1}.}
}
\value{
Generated plots are saved to \verb{output > 2_Normalization > peaks} directory. Generated text file is saved to \verb{output > 2_Normalization} directory.
}
\description{
This function generates one text file per parameter which contains all the determined peaks for every sample.
}
