% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{applyClusterModel}
\alias{applyClusterModel}
\title{Apply the training clustering on the whole dataset}
\usage{
applyClusterModel(
  dataTraining = dataTraining,
  dataValidation = dataValidation,
  parametersToUse = NULL,
  coresNumber = 1,
  chunksMaxSize = 1e+05
)
}
\arguments{
\item{dataTraining}{A data frame containing the training dataset. Defaults to \code{NULL}.}

\item{dataValidation}{A data frame containing the validation dataset. Defaults to \code{NULL}.}

\item{parametersToUse}{An vector defining the parameters to use for the model application. Defaults to \code{NULL}.}

\item{coresNumber}{An integer defining the number of cores to use to apply the clustering model. Defaults to \code{1}.}

\item{chunksMaxSize}{An integer defining the maximum cell number of a chunk to be processed. Defaults to \code{100000}.}
}
\value{
A data frame containing the clustered cells of the \code{dataValidation} dataset.
}
\description{
This function allows to apply the previously generated clustering on the remaining dataset not used in the training (also called validation dataset).
}
