% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{bindData}
\alias{bindData}
\title{Bind data to a single data frame}
\usage{
bindData(data = NULL)
}
\arguments{
\item{data}{A list containing the rearranged dataset. This argument should be the output of the \code{hierarchicalClusteringData()}. Defaults to \code{NULL}.}
}
\value{
A data frame containing the whole columns from \code{data}.
}
\description{
This function allows transform the rearranged dataset to simple data frame, to make subsequent data exportation and analysis easier.
}
