% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Preprocessing.R
\name{checkNewVersion}
\alias{checkNewVersion}
\title{Check new version}
\usage{
checkNewVersion(...)
}
\arguments{
\item{...}{Useful to transfer other arguments to the function.}
}
\description{
This function checks whether there is a new available version for \code{PICAFlow} in the GitHub repository. If this is the case, users are notified.
}
