% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{clustersPhenotypesHeatmap}
\alias{clustersPhenotypesHeatmap}
\title{Export information about clusters phenotype}
\usage{
clustersPhenotypesHeatmap(
  prefix = NULL,
  thresholds = NULL,
  metricUsed = "median",
  datasetFolder = "full"
)
}
\arguments{
\item{prefix}{A string defining the prefix of the clusters phenotypes file to use. Defaults to \code{NULL}.}

\item{thresholds}{A data frame defining the thresholds for each parameter. Defaults to \code{NULL}.}

\item{metricUsed}{A string defining the metric to use for statistics computation. It can be either \code{mean} or \code{median}. Defaults to \code{median}.}

\item{datasetFolder}{A string defining in which folder to save the results. It can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the data used. Defaults to \code{full}.}
}
\value{
Generated PDF files are saved to \verb{output > 7_Clustering > datasetFolder} directory. Generated text files are saved to \verb{output > 7_Clustering > datasetFolder} directory.
}
\description{
This function allows to export heatmaps and text files about clusters phenotype. It also summarizes the phenotypes by using binary mapping of the data using the pre-determined thresholds for each parameter.
}
