% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_UMAP.R
\name{determineOptimalUMAPParameters}
\alias{determineOptimalUMAPParameters}
\title{Determine optimal UMAP parameters}
\usage{
determineOptimalUMAPParameters(
  data = NULL,
  nNeighborsToTest = NULL,
  minDistToTest = NULL,
  downsample = 50000,
  datasetFolder = "full"
)
}
\arguments{
\item{data}{A data frame on which to perform the analysis. It must originate from \code{openDownsampledData()} method call. Defaults to \code{NULL}.}

\item{nNeighborsToTest}{A numeric vector defining the different \code{n_neighbors} to test. Defaults to \code{NULL}.}

\item{minDistToTest}{A numeric vector defining the different \code{min_dist} to test. Defaults to \code{NULL}.}

\item{downsample}{An integer defining the maximum number of cells to use for the analysis. Defaults to \code{50000}.}

\item{datasetFolder}{A string defining which dataset to open. Can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the \code{data} argument. Defaults to \code{full}.}
}
\value{
Generated PDF file is saved to \verb{output > 5_UMAP} directory.
}
\description{
This function computes UMAP dimensionality reduction using different combinations of \code{n_neighbors} and \code{min_dist} parameters to allow the user to choose the combination that best suits the current dataset. Usually, this function is called on a small subdataset, typically the cells that are tagged as downsampled.
}
