% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{exportClustersStatsAndPlots}
\alias{exportClustersStatsAndPlots}
\title{Export final clusters statistics and plots}
\usage{
exportClustersStatsAndPlots(
  data = NULL,
  folder = NULL,
  parametersToUse = NULL,
  coresNumber = 2,
  prefix = NULL,
  maxCellsPerPlot = 1000,
  metricUsed = "median",
  datasetFolder = "full"
)
}
\arguments{
\item{data}{A data frame containing the clustered dataset. It should be the output of \code{openClusteredFullDataCollapsed()} function. Defaults to \code{NULL}.}

\item{folder}{A string defining the directory in which the cluster plots will be saved. Defaults to \code{NULL}.}

\item{parametersToUse}{A character vector defining the parameters to use in the exported statistics. Defaults to \code{NULL}.}

\item{coresNumber}{An integer defining the number of processor cores to be used for data exportation. Defaults to \code{2}.}

\item{prefix}{A string defining the prefix to adjunct to each cluster name in each plot exported. Defaults to \code{NULL}.}

\item{maxCellsPerPlot}{An integer defining the maximum cell number displayed per plot. Defaults to \code{1000}.}

\item{metricUsed}{A string defining the metric to use for statistics computation. It can be either \code{mean} or \code{median}. Defaults to \code{median}.}

\item{datasetFolder}{A string defining in which folder to save the results. It can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the data used. Defaults to \code{full}.}
}
\value{
Generated PDF plots are saved to \verb{output > 7_Clustering > datasetFolder > folder} directory. Generated text files are saved to \verb{output > 7_Clustering > datasetFolder} directory.
}
\description{
This function allows to export statistics and plots for the final determined clusters.
}
