% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{exportDataBind}
\alias{exportDataBind}
\title{Export data bind}
\usage{
exportDataBind(data = NULL)
}
\arguments{
\item{data}{A data frame containing the dataset to export. This argument should be the output of the \code{bindData()} function. Defaults to \code{NULL}.}
}
\value{
Generated text file is saved to \verb{output > 8_Analysis} directory.
}
\description{
This function allows to export the final data bind to a tabular text file.
}
