% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Downsampling.R
\name{exportDownsamplingOutput}
\alias{exportDownsamplingOutput}
\title{Export all outputs from downsampling}
\usage{
exportDownsamplingOutput(poolData = NULL)
}
\arguments{
\item{poolData}{A list of length 2 produced by \code{poolData()} method. The \code{data} element will be exported as a new \code{rds} file, whereas the \code{message} element will be exported as a text file. Defaults to \code{NULL}.}
}
\value{
Generated \code{rds} file is saved to \code{rds} directory. Generated text file is saved to \verb{output > 4_Downsampling} directory.
}
\description{
This function allows to export data and log from the downsampling step.
}
