% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_FCS.R
\name{exportFCS}
\alias{exportFCS}
\title{Export FCS files from flowSet}
\usage{
exportFCS(data = NULL, datasetFolder = "full")
}
\arguments{
\item{data}{A data frame from which to generate the FCS files. It must originate from \code{openDownsampledData()} method call and must contain the UMAP parameters for each cell. Defaults to \code{NULL}.}

\item{datasetFolder}{A string defining in which folder to save the results. It should match the \code{datasetToUse} value. Can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the \code{data} argument. Defaults to \code{full}.}
}
\value{
Generated FCS and text files are saved to \verb{output > 6_FCS} directory.
}
\description{
This function allows to export new FCS files from a data frame.
}
