% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{exportFeaturesTables}
\alias{exportFeaturesTables}
\title{Export global tables for defined features}
\usage{
exportFeaturesTables(
  data = NULL,
  columnsToUse = NULL,
  featuresToUse = NULL,
  metricToUse = "median"
)
}
\arguments{
\item{data}{A data frame containing the dataset to export. This argument should be the output of the \code{bindData()} function. Defaults to \code{NULL}.}

\item{columnsToUse}{A character vector defining the columns considered as the variable of interest (typically cell clusters). Defaults to \code{NULL}.}

\item{featuresToUse}{A character vector defining the columns to be used as features (typically group, cluster or metadata). Defaults to \code{NULL}.}

\item{metricToUse}{A string defining the metric to use for statistics computation. It can be either \code{mean} or \code{median}. Defaults to \code{median}.}
}
\value{
Generated text files are saved to \verb{output > 8_Analysis} directory.
}
\description{
This function allows to export text files which contain several statistics about columns of interest using selected features of interest as data delimiters.
}
