% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Gating.R
\name{exportGatingParameters}
\alias{exportGatingParameters}
\title{Export gating parameters}
\usage{
exportGatingParameters(gatingParameters = NULL)
}
\arguments{
\item{gatingParameters}{A list defining the parameters to export. It should be a list generated using \code{gateData()} function. Defaults to \code{NULL}.}
}
\value{
Generated text file is saved to \verb{output > 3_Gating} directory.
}
\description{
This function allows to export a text file summarizing all the parameters used for the gating process.
}
