% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{exportParametersUsed}
\alias{exportParametersUsed}
\title{Export all parameters used}
\usage{
exportParametersUsed(parametersToExport = NULL)
}
\arguments{
\item{parametersToExport}{A list containing all the parameters to export. Defaults to \code{NULL}.}
}
\value{
Generated text file is saved to \code{output} directory.
}
\description{
This function allows to export all the parameters that were used throughout the entire analysis. Some parameters will not be exported here because they were already exported in a special format when the associated functions were used.
}
