% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Transformation.R
\name{exportPerParameter}
\alias{exportPerParameter}
\title{Export a single \code{rds} file per parameter}
\usage{
exportPerParameter(parametersToExport = NULL, nCoresToExploit = NULL)
}
\arguments{
\item{parametersToExport}{A vector of all parameters to export. Defaults to \code{NULL}.}

\item{nCoresToExploit}{An integer defining the number of cores to use to export the \code{rds} files. Defaults to \code{NULL} which translates to the actual number of cores of the processor minus 1.}
}
\value{
Generated \code{rds} files are saved to \code{rds} directory.
}
\description{
This function reconstructs the dataset by exporting a single \code{rds} file per parameter instead of one \code{rds} file per sample. Please note that after the completion of the exportation, the sample-specific \code{rds} file will be deleted.
}
