% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Normalization.R
\name{exportRDSPeaksDataToTabularTextFiles}
\alias{exportRDSPeaksDataToTabularTextFiles}
\title{Reexport peak tabular text files from RDS peaks data structure}
\usage{
exportRDSPeaksDataToTabularTextFiles()
}
\description{
This function allows to reexport all the tabular text files containing all the corrected peaks for every parameter and sample after peaks have been corrected through the dedicatedShiny application.
}
