% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{FastPhenoGraph_clustering}
\alias{FastPhenoGraph_clustering}
\title{Perform PhenoGraph clustering}
\usage{
FastPhenoGraph_clustering(
  data = NULL,
  parametersToUse = NULL,
  k = 100,
  coresNumber = 1,
  datasetFolder = "full"
)
}
\arguments{
\item{data}{A data frame containing the dataset to use. Defaults to \code{NULL}.}

\item{parametersToUse}{A vector defining the parameters to use for the clustering. Defaults to \code{NULL}.}

\item{k}{An integer defining the maximum number of clusters to generate. PhenoGraph will surely output way less than this number, depending on the dataset of interest. Defaults to \code{100}.}

\item{coresNumber}{An integer defining the number of cores to use to apply the clustering model. Defaults to \code{1}.}

\item{datasetFolder}{A string defining in which folder to save the results. It can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the data used. Defaults to \code{full}.}
}
\value{
Generated rds file is saved to \code{rds} directory.
}
\description{
This function allows to perform PhenoGraph clustering algorithm on the dataset and parameters of interest.
}
