% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Normalization.R
\name{fineTunePeaks}
\alias{fineTunePeaks}
\title{Manually fine tune peaks}
\usage{
fineTunePeaks()
}
\description{
This function allows to manually fine tune the pre-determined peaks for each sample and parameter using a dedicated Shiny application.
}
