% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Gating.R
\name{gateData}
\alias{gateData}
\title{Gate on cells of interest}
\usage{
gateData(
  flowset = NULL,
  sampleToPlot = NULL,
  xParameter = NULL,
  yParameter = NULL,
  xlim = NULL,
  ylim = NULL,
  subset = FALSE,
  gateName = NULL,
  exportAllPlots = FALSE,
  samplesPerPage = 6,
  recursivity = FALSE,
  inverseGating = FALSE,
  specificGatesSampleIDs = NULL,
  redrawGate = TRUE,
  gatingset = NULL,
  generatedGates = NULL,
  customBinWidth = 2000
)
}
\arguments{
\item{flowset}{A string defining the \code{flowSet} to use. Defaults to \code{NULL}.}

\item{sampleToPlot}{A numeric vector defining which samples should be plotted for gating visualization. Defaults to \code{NULL}.}

\item{xParameter}{A string defining which parameter to use as x. Defaults to \code{NULL}.}

\item{yParameter}{A string defining which parameter to use as y. Defaults to \code{NULL}.}

\item{xlim}{A numeric vector of length 2 defining the minimum and maximum display values for the x axis. Defaults to \code{NULL}.}

\item{ylim}{A numeric vector of length 2 defining the minimum and maximum display values for the y axis. Defaults to \code{NULL}.}

\item{subset}{A boolean defining if gated cells should be actually extracted from the \code{flowSet}. Defaults to \code{FALSE}.}

\item{gateName}{A string defining a user-friendly name for the current gate. Defaults to \code{NULL}.}

\item{exportAllPlots}{A boolean defining if generated plots should be exported as PDF. If set to \code{TRUE}, all samples will be plotted according to the \code{samplesPerPage} argument. Defaults to \code{FALSE}.}

\item{samplesPerPage}{An integer defining the number of samples to be plotted per PDF page. This argument is only used when \code{exportAllPlots = TRUE}. Defaults to \code{6}.}

\item{recursivity}{A boolean defining if generated plots should be simply printed (if set to \code{FALSE}) or returned by the function (if set to \code{TRUE}). This argument is automatically set to \code{TRUE} when \code{exportAllPlots} is also set to \code{TRUE}. Please note that this argument should not be changed manually. Defaults to \code{FALSE}.}

\item{inverseGating}{A boolean defining if the gate should rather consider cells that are outside the gate (if set to \code{TRUE}). Defaults to \code{FALSE}.}

\item{specificGatesSampleIDs}{A numeric vector defining the sample IDs to use for the generation of sample-wise adapted gates. Defaults to \code{NULL}}

\item{redrawGate}{A boolean defining if the gates should be manualy redrawn or not. Defaults to \code{TRUE}}

\item{gatingset}{A \code{gatingset} object used internally to store data in a compatible way with \code{flowGate} package. This value should not be manually provided apart from what is described in the tutorial. Defaults to \code{NULL}.}

\item{generatedGates}{A named list containing the gates that were generated during the gating process. This value should not be manually provided apart from what is described in the tutorial. Defaults to \code{NULL}.}

\item{customBinWidth}{A numeric vector defining the width of the bins to use for data plotting during interactive gating (within the Shiny application) and after interactive gating (in the exported PDF files). High values (typically over \code{100} or \code{1000}, sometimes higher than that) are better suited for linear parameters with very broad range (such FSC, SSC, etc.) whereas low values (typically under \code{10} or \code{1}, sometimes lower than that) are better suited for log-transformed parameters with a small range. We highly recommend to try several values chosen from both ranges to select the one that best fits your dataset. Defaults to \code{2000}.}
}
\value{
If \code{exportAllPlots = TRUE}, the function will output PDF files containing the gating plots for each sample of the dataset. If \code{subset = TRUE}, the function will return a list of 2 elements named \code{flowset} (containing the actual gated \code{flowSet}) and \code{summary} (containing basic statistics about the gating such as the number of cells before gating, the number of cells gated and the proportion of cells gated).
}
\description{
This function allows to gate on cells of interest.
}
