% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Preprocessing.R
\name{getAllChannelsInformation}
\alias{getAllChannelsInformation}
\title{Get all channels information}
\usage{
getAllChannelsInformation()
}
\value{
A dataframe containing the ID, name and description of all the found parameters.
}
\description{
This function gets the names and descriptions of all the parameters contained in the first available \code{rds} file in the \code{rds} directory.
}
