% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Gating.R
\name{getParameterLimits}
\alias{getParameterLimits}
\title{Get 2nd and 99th percentiles from parameter distribution}
\usage{
getParameterLimits(flowset = NULL, sample = 1, parameter = NULL)
}
\arguments{
\item{flowset}{A string defining the \code{flowSet} to use. Defaults to \code{NULL}.}

\item{sample}{A numeric vector defining which sample should be used to get the limits. Defaults to \code{1}.}

\item{parameter}{A string defining which parameter to extract limits from. Defaults to \code{NULL}.}
}
\value{
A vector of length 2 containing the 2nd and 99th percentiles, in the first and second position, respectively.
}
\description{
This function allows to extract the 2nd and 99th percentiles for any given parameter distribution contained in the dataset.
}
