% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{heatmapAbundancesGroups}
\alias{heatmapAbundancesGroups}
\title{Export clusters abundance heatmap using a given feature of interest}
\usage{
heatmapAbundancesGroups(
  feature = NULL,
  clustersToKeepRegex = NULL,
  metricToUse = "median"
)
}
\arguments{
\item{feature}{A string defining the column to be used a feature (typically group, cluster or metadata). Defaults to \code{NULL}.}

\item{clustersToKeepRegex}{A string defining the regex to use for the identification of columns associated to cell clusters. Defaults to \code{NULL}.}

\item{metricToUse}{A string defining the metric to use for statistics computation. It can be either \code{mean} or \code{median}. Defaults to \code{median}.}
}
\value{
Generated PDF file is saved to \verb{output > 8_Analysis} directory.
}
\description{
This function allows to export a heatmap showing clusters abundance using a given feature of interest as delimiter.
}
