% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{hierarchicalClusteringData}
\alias{hierarchicalClusteringData}
\title{Perform hierarchical clustering of UMAP parameters}
\usage{
hierarchicalClusteringData(data = NULL, clustersNb = NULL)
}
\arguments{
\item{data}{A list containing the rearranged dataset. This argument should be the output of the \code{UMAP_clusters()} function with \code{addUMAPOutputToData = TRUE}. Defaults to \code{NULL}.}

\item{clustersNb}{An integer defining the number expected/observable/supposed clusters from the UMAP plot. Defaults to \code{NULL}.}
}
\value{
A list of length 2: the \code{dataSubset} element containing a data frame of the desired columns of interest and the \code{dataRemoved} element containing the remaining columns of the dataset.
}
\description{
This function allows to perform hierarchical clustering of UMAP parameters. Please note that the determined clusters for each cell will be saved as a new column named \code{cluster} in the \code{dataRemoved} element of \code{data}.
}
