% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Normalization.R
\name{keepPreviousPeaksData}
\alias{keepPreviousPeaksData}
\title{Generate new peak files using previously generated peak files}
\usage{
keepPreviousPeaksData()
}
\description{
This function generates new peak files from previously generated peak files but keeps the data already present in the old peak files. Typically, this function is used when new samples are added to the dataset but peak information from the already present samples should be kept. To use this function, please create a \code{refs} directory within the \verb{output > 2_Normalization > peaks} directory, then copy the old peak files within this \code{refs} directory. Afterwards, generate the peak files for the current dataset using \code{analyzePeaks()} function, then run the \code{keepPreviousPeaksData()} function. Newly generated peak files will overwrite the ones present in the \verb{output > 2_Normalization > peaks} directory. Basically, if a sample is found in both new and old peak files, corresponding data from old peak files will overwrite the corresponding data from new peak files, thus keeping intact the already determined information in the old dataset. If needed, a backup (before edition) of the peak files located in the \verb{output > 2_Normalization > peaks} can be found in the \verb{output > 2_Normalization > peaks > backup} directory.
}
