% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Transformation.R
\name{launchManualLogicleShinyApp}
\alias{launchManualLogicleShinyApp}
\title{Launch the R Shiny interactive application for visualization of data logicle transformation}
\usage{
launchManualLogicleShinyApp(fs_shiny = NULL)
}
\arguments{
\item{fs_shiny}{A flowSet generated using \code{mergeData()} function. Defaults to \code{NULL}.}
}
\description{
This function launches a R Shiny interactive application which helps to visually determine and adjust the parameters for logicle transformation of data.
}
