% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{mergeData}
\alias{mergeData}
\title{Open and merge clusters abundance files}
\usage{
mergeData(pattern = NULL, sampleNamesColumn = NULL)
}
\arguments{
\item{pattern}{A string defining the pattern used to determine which files should be opened. Defaults to \code{NULL}.}

\item{sampleNamesColumn}{A string defining the column name the be used as common comparator for each file to merge. This column should be present in each file to be opened and should present identical values (for instance, individualized sample names). Defaults to \code{NULL}.}
}
\value{
A data frame containing the merged data.
}
\description{
This function allows to open and merge one or several clusters abundance files. Please note that the input text file(s) should be generated with the \code{exportClustersStatsAndPlots()} function and placed in the \verb{output > 8_Analysis} directory.
}
