% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{mergeMetadata}
\alias{mergeMetadata}
\title{Open and merge metadata to dataset}
\usage{
mergeMetadata(data = NULL, mergeColumn = NULL, replaceColumn = NULL)
}
\arguments{
\item{data}{A data frame containing the original dataset to which metadata are to be applied. This argument should be the output of the \code{mergeData()} function. Defaults to \code{NULL}.}

\item{mergeColumn}{A string defining the column name the be used as comparator for the metadata. This column should be present in the metadata file and contain values that match the rownames of \code{data} (for instance, individualized sample names). Defaults to \code{NULL}.}

\item{replaceColumn}{A string defining the column name the be used as the final name for samples. This column should be present in the metadata file. Defaults to \code{NULL}.}
}
\value{
A data frame containing the merged data.
}
\description{
This function allows to open and merge metadata (such as clinical data) to the dataset. Please note that the metadata file should be an Excel spreadsheet named \code{metadata.xslx} and placed in the \verb{output > 8_Analysis} directory.
}
