% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Normalization.R
\name{mergeParameters}
\alias{mergeParameters}
\title{Merge all parameters into a single file}
\usage{
mergeParameters(suffix = "_normalized")
}
\arguments{
\item{suffix}{A string defining the \code{rds} files to use. Defaults to \verb{_normalized}.}
}
\value{
Generated \code{rds} file is saved to \code{rds} directory.
}
\description{
This function merges all the parameter-restricted \code{rds} files into a single \code{flowSet} contained in a \code{rds} file. Please note that once the merging process finished, input \code{rds} files will be deleted.
}
