% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Normalization.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize peaks for each sample and parameter}
\usage{
normalizeData(
  try = TRUE,
  max.lms.sequence = NULL,
  suffix = NULL,
  base.lms.list = NULL,
  warpSet = FALSE,
  gaussNorm = TRUE,
  samplesToDelete = NULL,
  nCoresToExploit = NULL,
  custom.lms.list = NULL
)
}
\arguments{
\item{try}{A boolean defining if the normalization should be performed for real or as a test. Defaults to \code{TRUE}.}

\item{max.lms.sequence}{A named numeric list containing the maximum expected/desired number of peaks for each parameter. Defaults to \code{NULL}.}

\item{suffix}{A string defining the \code{rds} files to use. Defaults to \verb{_raw}.}

\item{base.lms.list}{A list containing the target peaks values to reach. Should be the \code{best} list element from the \code{synthesizePeaks()} method output. Please note that this argument is only used when the normalization method is set to \code{gaussNorm}.}

\item{warpSet}{A boolean defining if the normalization should use the \code{warpSet()} method from \code{flowStats} package. Defaults to \code{FALSE}.}

\item{gaussNorm}{A boolean defining if the normalization should use the \code{gaussNorm2()} method adapted from \code{flowStats} package. Defaults to \code{TRUE}.}

\item{samplesToDelete}{A character vector defining the samples to remove before proceeding to the normalization. Defaults to \code{NULL}.}

\item{nCoresToExploit}{An integer defining the number of cores to use to export the \code{rds} files. Defaults to \code{NULL} which translates to the actual number of cores of the processor minus 1.}

\item{custom.lms.list}{A list containing the actual peaks values for each sample and parameter. Should be the \code{raw} list element from the \code{synthesizePeaks()} method output. Please note that this argument is only used when the normalization method is set to \code{gaussNorm}. Defaults to \code{NULL}.}
}
\value{
If \code{try = TRUE}, the function will only return the output messages from the normalization attempt. If \code{try = FALSE}, generated \code{rds} files are saved to \code{rds} directory.
}
\description{
This function normalizes peaks for each sample and parameter, using a slightly modified version of the \code{gaussNorm()} method from the \code{flowStats} package or the unmodified \code{warpSet()} method from \code{flowStats} package. This version allows to take the peaks individually determined for each sample into account, in order to make them reach the target values for each peak of each parameter.
}
