% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_UMAP.R
\name{openDownsampledData}
\alias{openDownsampledData}
\title{Open the full, training or validation downsampled dataset}
\usage{
openDownsampledData(datasetToUse = "full")
}
\arguments{
\item{datasetToUse}{A string defining which dataset to open. Can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. Defaults to \code{full}.}
}
\value{
Returns the associated \code{rds} data.
}
\description{
This function allows to open one of the three possible downsampled datasets: the full dataset, the training subdataset or the validation subdataset.
}
