% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Normalization.R
\name{plotFacets}
\alias{plotFacets}
\title{Plot densities for each parameter}
\usage{
plotFacets(
  parametersToPlot = NULL,
  maxSamplesNbPerPage = 10,
  folder = NULL,
  suffix = "_raw",
  downsample_n = 50000,
  nCoresToExploit = NULL,
  plotPeaks = FALSE
)
}
\arguments{
\item{parametersToPlot}{A vector of all parameters to plot. Defaults to \code{NULL}.}

\item{maxSamplesNbPerPage}{An integer defining the maximum number of samples to plot on each page. Defaults to \code{10}.}

\item{folder}{A string defining the folder to create for the current graphs generation. Defaults to \code{NULL}.}

\item{suffix}{A string defining the pattern of which \code{rds} files to use. Defaults to \verb{_raw}.}

\item{downsample_n}{An integer defining the maximum number of cells to use for each plotted sample. Defaults to \code{50000}.}

\item{nCoresToExploit}{An integer defining the number of cores to use to export the graphs. It corresponds to the number of parameters to be treated concomitantly. Defaults to \code{NULL}, which translates to the actual number of cores of the processor minus 1.}

\item{plotPeaks}{A boolean defining if the peaks values found in the corresponding tabular text files should be displayed on each graph (if set to \code{TRUE}) or not (if set to \code{FALSE}).}
}
\value{
Generated plots are saved to \verb{output > 2_Normalization > folder} directory.
}
\description{
This function plots, for each parameter, the stacked densities of fluorescence of several samples on a single page.
}
