% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{plotFinalClusteringTraining}
\alias{plotFinalClusteringTraining}
\title{Export plot showing the cutoff on the previous decision plot}
\usage{
plotFinalClusteringTraining(
  initialClustering = NULL,
  finalClustering = NULL,
  datasetFolder = "full"
)
}
\arguments{
\item{initialClustering}{An object generated using the \code{initialClusteringTraining()} function. Defaults to \code{NULL}.}

\item{finalClustering}{An object generated using the \code{finalClusteringTraining()} function. Defaults to \code{NULL}.}

\item{datasetFolder}{A string defining in which folder to save the results. It can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the data used. Defaults to \code{full}.}
}
\value{
A \code{ggplotly} plot. Please note that after the cutoff selection is performed, the graph should be close with \code{dev.off()} call.
}
\description{
This function allows to export the previous decision plot on which is displayed the chosen cutoff for hierarchical clustering tree cutting.
}
