% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{plotUMAP_projectionFinalClusters}
\alias{plotUMAP_projectionFinalClusters}
\title{Export UMAP plot with final clusters}
\usage{
plotUMAP_projectionFinalClusters(
  data = NULL,
  displayedCells = 1e+05,
  datasetFolder = "full"
)
}
\arguments{
\item{data}{A data frame containing the clustered dataset. It should be the output of \code{openClusteredFullDataCollapsed()} function. Defaults to \code{NULL}.}

\item{displayedCells}{An integer defining the maximum number of cells to display on the plot. Defaults to \code{100000}.}

\item{datasetFolder}{A string defining in which folder to save the results. It can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the data used. Defaults to \code{full}.}
}
\value{
Generated PDF plot is saved to \verb{output > 7_Clustering} directory.
}
\description{
This function allows to export a PDF file showing the UMAP parameters of the involved dataset and their colored determined clusters.
}
