% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{plotUMAP_projectionTraining}
\alias{plotUMAP_projectionTraining}
\title{Export UMAP plot with clusters determined in training dataset}
\usage{
plotUMAP_projectionTraining(
  dataTrainingClustered = NULL,
  datasetFolder = "full"
)
}
\arguments{
\item{dataTrainingClustered}{A data frame containing the clustered training dataset. Defaults to \code{NULL}.}

\item{datasetFolder}{A string defining in which folder to save the results. It can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the data used. Defaults to \code{full}.}
}
\value{
Generated PDF plot is saved in the \verb{output > 7_Clustering} directory.
}
\description{
This function allows to export a PDF file showing the UMAP parameters of the involved training dataset and their colored determined clusters.
}
