% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{removeOutliers}
\alias{removeOutliers}
\title{Remove outliers from dataset}
\usage{
removeOutliers(data = NULL, outliers = NULL)
}
\arguments{
\item{data}{A list containing the merged dataset. This argument should be the output of the \code{mergeMetadata()} function. Defaults to \code{NULL}.}

\item{outliers}{An characted vector defining the outliers to remove from the dataset. Defaults to \code{NULL}.}
}
\value{
A data frame containing \code{data} minus the outliers.
}
\description{
This function allows to remove from the dataset samples that are considered as outliers. Please note that these samples will be permanently removed and cannot be recovered afterwards.
}
