% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{ROCanalysis}
\alias{ROCanalysis}
\title{Perform ROC analysis}
\usage{
ROCanalysis(
  dataFile = NULL,
  predictors = NULL,
  pairsToAnalyze = NULL,
  pairs_columnToCheck = NULL
)
}
\arguments{
\item{dataFile}{A string defining the Excel file to use. Defaults to \code{NULL}.}

\item{predictors}{A list defining the predictors to use. Each element of the list is considered as a predictor. If an element of the \code{predictors} list is itself a list, the final associated predictor will be the sum of the elements inside this list. Defaults to \code{NULL}.}

\item{pairsToAnalyze}{A list defining the pairs to use for the ROC analysis. Each element of the \code{pairsToAnalyze} list should also be a list of length 2 where both sub-elements are a vector representing the paired groups. Each of these sub-elements can be a vector of the desired length. In the case the vector length of a sub-element is greater than 1, the final group will consist of the pool of every sub-element put inside. Defaults to \code{NULL}.}

\item{pairs_columnToCheck}{A string defining the column to use for pairs identification. Defaults to \code{NULL}.}
}
\value{
Generated text and PDF files are saved to \verb{output > 8_Analysis > ROC} directory.
}
\description{
This function allows to perform ROC analysis on the full data exported using the \code{exportDataBind()} function. To use this function, please create a Excel copy (\code{xslx} format) of the \code{fullData.txt} file. One can use this Excel file to easily create new features, notably by combining several other ones (for instance by summing clusters).
}
