% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Preprocessing.R
\name{setupWorkingDirectory}
\alias{setupWorkingDirectory}
\title{Setup the working directory}
\usage{
setupWorkingDirectory()
}
\description{
This function setups the working directory. It does not take any input argument but assumes that the \code{workingDirectory} variable is set to the path of interest using the base R \code{setwd()} function. From this path, everything will be wiped and \code{input}, \code{output} and \code{rds} directories will be created. The `output`` directory will also contain several other specific directories that will be used in the subsequent analyses.
}
