% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Gating.R
\name{shinyAppGating}
\alias{shinyAppGating}
\title{Shiny application to gate on cells of interest}
\usage{
shinyAppGating(
  flowset = NULL,
  sample = 1,
  param_x = NULL,
  param_y = NULL,
  gateKind = "globalGate",
  param_x_minLim = NULL,
  param_x_maxLim = NULL,
  param_y_minLim = NULL,
  param_y_maxLim = NULL,
  bins = 5000
)
}
\arguments{
\item{flowset}{A string defining the \code{flowSet} to use. Defaults to \code{NULL}.}

\item{sample}{A numeric vector defining which sample should be plotted for gating visualization. Defaults to \code{1}.}

\item{param_x}{A string defining which parameter to use as x. Defaults to \code{NULL}.}

\item{param_y}{A string defining which parameter to use as y. Defaults to \code{NULL}.}

\item{gateKind}{A string defining the kind of gate to define (either \code{globalGate} or \code{specialGate}). Defaults to \code{globalGate}.}

\item{param_x_minLim}{A numeric defining the minimum graph limit to use for the parameter \code{param_x} on the x-axis. Defaults to \code{NULL}.}

\item{param_x_maxLim}{A numeric defining the maximum graph limit to use for the parameter \code{param_x} on the x-axis. Defaults to \code{NULL}.}

\item{param_y_minLim}{A numeric defining the maximum graph limit to use for the parameter \code{param_y} on the y-axis. Defaults to \code{NULL}.}

\item{param_y_maxLim}{A numeric defining the maximum graph limit to use for the parameter \code{param_y} on the y-axis. Defaults to \code{NULL}.}

\item{bins}{A numeric vector defining the width of the bins to use for data plotting after interactive gating. Defaults to \code{5000} (value passed by the mother \code{gateData()} function). High values (over \code{100} or \code{1000}) are better suited for linear parameters with very broad range (such FSC, SSC, etc.) whereas low values (under \code{10} or \code{1}) are better suited for log-transformed parameters with a small range.}
}
\value{
The function will not directly return any value. Instead, it will write a RDS file containing the necessary gate information which can be loaded and used afterwards.
}
\description{
This Shiny application allows to interactively gate on cells of interest. Warning: this function is called when using the \code{gateData()} mother function, so it should not be called directly.
}
