% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Downsampling.R
\name{splitDataset}
\alias{splitDataset}
\title{Generate two subdatasets from downsampled dataset}
\usage{
splitDataset(trainingDatasetProportion = 0.75)
}
\arguments{
\item{trainingDatasetProportion}{An decimal numeric value (between 0 and 1) defining the frequency of cells that will go to the training subdataset. Defaults to \code{0.75}.}
}
\value{
Generated \code{rds} files are saved to \code{rds} directory.
}
\description{
This function allows to export two different subdatasets from the original downsampled dataset. This is typically used when one wants to generate training and validation datasets. Please note that this function do not directly split the samples between the two subdatasets, but rather the cells themselves. It means that at the end, each subdataset will have the same number of samples as compared to the original dataset, but will not contain the same cells: some will be attributed to the training subdataset, whereas the others will be attributed to the validation subdataset.
}
