% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Preprocessing.R
\name{subsetData}
\alias{subsetData}
\title{Subset \code{rds} files}
\usage{
subsetData(parametersToKeep = NULL, customNames = NULL)
}
\arguments{
\item{parametersToKeep}{A vector of all parameters to keep in each \code{rds} file. Defaults to \code{NULL}.}

\item{customNames}{A vector of the same length as \code{parametersToKeep} containing the user-friendly parameter names. Defaults to \code{NULL}.}
}
\value{
Generated \code{rds} files are saved to \code{rds} directory, overwriting the previous ones.
}
\description{
This function subsets every \code{rds} file with the specified parameters of interest. It can also rename each parameter in a more user-friendly way.
}
