% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{subsetDataUMAP}
\alias{subsetDataUMAP}
\title{Construct rearranged datasets ready for further analyses}
\usage{
subsetDataUMAP(
  data = NULL,
  columnsToKeepData = NULL,
  columnsToKeepMetadata = NULL,
  isColumnsToKeepDataRegex = TRUE
)
}
\arguments{
\item{data}{A data frame containing the original dataset to rearrange. This argument should be the output of the \code{mergeMetadata()} function. Defaults to \code{NULL}.}

\item{columnsToKeepData}{A character vector defining the columns of the dataset to keep. This argument was rather designed for actual clusters data, but can be used for other kind of columns. Defaults to \code{NULL}. Please note that this argument can either be considered as a string (if \code{isColumnsToKeepDataRegex = FALSE}) or as a regex (if \code{isColumnsToKeepDataRegex = TRUE}).}

\item{columnsToKeepMetadata}{A character vector defining other columns of the dataset to keep. This argument was rather designed for metadata, but can be used for other kind of columns. Defaults to \code{NULL}. Please note that this argument can only be considered as a string but never as a regex.}

\item{isColumnsToKeepDataRegex}{A boolean defining if the columnsToKeepData argument should be considered as a string (if \code{isColumnsToKeepDataRegex = FALSE}) or as a regex (if \code{isColumnsToKeepDataRegex = TRUE}). Defaults to \code{TRUE}.}
}
\value{
A list of length 2: the \code{dataSubset} element which contains a data frame of the desired columns of interest and the \code{dataRemoved} element which contains the remaining columns of the dataset.
}
\description{
This function allows to subset and rearrange a dataset file for subsequent analyses. If needed, desired column can also be extracted. Please note that no data are removed from the dataset, but are rather rearranged. Columns of interest will be put in the \code{dataSubset} element of the final output list, whereas other columns will be put in the \code{dataRemoved} element of the final output list.
}
