% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_Normalization.R
\name{synthesizePeaks}
\alias{synthesizePeaks}
\title{Synthesize peaks information into a single object}
\usage{
synthesizePeaks()
}
\value{
A named list of size 3 (\code{info}, \code{best} and \code{raw}) containing all the summarized elements of the peak files. The \code{info} element contains general information about the peaks, organized as a table with the following columns: \code{Parameter} (parameter name), \code{PeaksNb} (number of peaks specified), \code{PeakType} (which precise peak is described) and \code{Mean} (the mean of all the peaks for every sample). The \code{best} element is a named list containing the mean peaks found for each parameter. The \code{raw} element stores the total peaks list for every parameter and sample.
}
\description{
This function takes all peak files from the \verb{output > 2_Normalization > peaks} directory and returns a single object which contains all the useful elements for the next steps.
}
