% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{testDatasetNormality}
\alias{testDatasetNormality}
\title{Test normality of dataset parameters}
\usage{
testDatasetNormality(
  data = NULL,
  parametersToAnalyze = NULL,
  datasetFolder = "full"
)
}
\arguments{
\item{data}{A string defining the dataset to use. Defaults to \code{NULL}.}

\item{parametersToAnalyze}{A character vector defining the parameters to analyze in the subsequent dataset. Defaults to \code{NULL}.}

\item{datasetFolder}{A string defining which dataset to open. It can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the data used. Defaults to \code{full}.}
}
\value{
Generated text file is saved to \verb{output > 7_Clustering > datasetFolder} directory.
}
\description{
This function allows to test the normality of each parameter of the dataset using Shapiro-Wilk test. This is useful to determine whether mean or median metric should be used in later analyses. Typically, a skewed distribution will show poorly informative mean values, but much more informative median values.
}
