% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Transformation.R
\name{transformData}
\alias{transformData}
\title{Transform \code{rds} files}
\usage{
transformData(parametersToTransform = NULL)
}
\arguments{
\item{parametersToTransform}{A vector of all parameters to transform in each \code{rds} file. Defaults to \code{NULL}.

Please note that this function needs the \code{parametersTransformations.rds} file located in the \verb{output > 1_Transformation} directory which is generated using the \code{transformationTuning} R Shiny application. This file contains the type as well as the parameters of the transformation of interest to be used for each flow cytometry parameter in the dataset.}
}
\value{
Generated \code{rds} files are saved to \code{rds} directory, overwriting the previous ones.
}
\description{
This function applies a mathematical transformation to data from each \code{rds} file.
}
