% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{UMAP_clusters}
\alias{UMAP_clusters}
\title{Perform UMAP analysis of columns of interest from specified dataset}
\usage{
UMAP_clusters(
  data = NULL,
  n_neighbors_UMAP = NULL,
  min_dist_UMAP = NULL,
  feature = NULL,
  suffix = NULL,
  returnUMAPData = FALSE,
  computeUMAP = TRUE,
  seedValue = 42,
  plotHighlightedFeatureItems = TRUE
)
}
\arguments{
\item{data}{A list containing the rearranged dataset. This argument should be the output of the \code{subsetDataUMAP()} function. Defaults to \code{NULL}.}

\item{n_neighbors_UMAP}{An integer defining the number of neighbors to use for UMAP analysis. Defaults to \code{NULL}.}

\item{min_dist_UMAP}{An integer defining the minimum distance to use for UMAP analysis. Defaults to \code{NULL}.}

\item{feature}{A string defining the feature used for overlay and ellipses construction to the UMAP plot. This feature should be a column name of the \code{dataRemoved} element of \code{data}. Defaults to \code{NULL}.}

\item{suffix}{A string defining a custom suffix to add to the plot name. Defaults to \code{NULL}.}

\item{returnUMAPData}{A boolean defining if UMAP parameters (\code{UMAP_1} and \code{UMAP_2}) should be added to the \code{dataRemoved} element of \code{data}. Defaults to \code{FALSE}.}

\item{computeUMAP}{A boolean defining if actual UMAP computation should be performed or not. Defaults to \code{TRUE}.}

\item{seedValue}{An integer defining the seed value to use for UMAP computation. Setting a seed helps to obtain reproducible UMAP plots. Defaults to \code{42}.}

\item{plotHighlightedFeatureItems}{A boolean defining if UMAP plots displaying highlighted items of the selected \code{feature} should also be generated or not. Defaults to \code{TRUE}.}
}
\value{
A list of length 2: the \code{dataSubset} element containing a data frame of the desired columns of interest and the \code{dataRemoved} element containing the remaining columns of the dataset. Generated PDF files are saved to \verb{output > 8_Analysis} directory.
}
\description{
This function allows to perform UMAP analysis of the columns of interest from specified dataset.
}
