% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_UMAP.R
\name{UMAP_downsampledDataset}
\alias{UMAP_downsampledDataset}
\title{Compute UMAP on the whole downsampled dataset}
\usage{
UMAP_downsampledDataset(
  data = NULL,
  n_threads = 1,
  min_dist = NULL,
  n_neighbors = NULL
)
}
\arguments{
\item{data}{A data frame on which to perform the analysis. It must originate from \code{openDownsampledData()} method call. Defaults to \code{NULL}.}

\item{n_threads}{An integer defining the number of cores to use for UMAP computation. Be careful, this function is expensive in terms of computational resources, so do not increase this value too much. Defaults to \code{1}.}

\item{min_dist}{A numeric vector defining the different \code{min_dist} to test. Defaults to \code{NULL}.}

\item{n_neighbors}{A numeric vector defining the different \code{n_neighbors} to test. Defaults to \code{NULL}.}
}
\value{
A full list of UMAP information computed for the provided data.
}
\description{
This function computes UMAP dimensionality reduction on the whole downsampled dataset and returns the associated UMAP embeddings as well as the model used.
}
