% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_UMAP.R
\name{UMAPFlowset}
\alias{UMAPFlowset}
\title{Apply UMAP model on the remaining cells}
\usage{
UMAPFlowset(data = NULL, model = NULL, chunksMaxSize = 1e+06, n_threads = 1)
}
\arguments{
\item{data}{A data frame on which to perform the analysis. It must originate from \code{openDownsampledData()} method call. Defaults to \code{NULL}.}

\item{model}{An object obtained after a call to umap() method from uwot package with ret_model argument set to TRUE. Defaults to \code{NULL}.}

\item{chunksMaxSize}{An integer defining the maximum cell number of each chunk for UMAP computation. Defaults to \code{1000000}.}

\item{n_threads}{An integer defining the number of cores to use for UMAP computation. Be careful, this function is expensive in terms of computational resources, so do not increase this value too much. Defaults to \code{1}.}
}
\value{
Returns the data and associated UMAP parameters.
}
\description{
This function allows to apply the UMAP model generated on downsampled cells to all the other cells. Usually, this function is called on a large subdataset, typically the cells that were not tagged as downsampled.
}
