% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_Clustering.R
\name{collapseCloseClusters}
\alias{collapseCloseClusters}
\title{Collapse the phenotypically close cell clusters}
\usage{
collapseCloseClusters(
  data = NULL,
  parametersToUse = NULL,
  parametersToUseThresholds = NULL,
  metricUsed = "median",
  datasetFolder = "full"
)
}
\arguments{
\item{data}{A data frame containing the dataset to use. Defaults to \code{NULL}.}

\item{parametersToUse}{A character vector defining the parameters to analyze in the subsequent dataset. Defaults to \code{NULL}.}

\item{parametersToUseThresholds}{A data frame defining the thresholds for each parameter Defaults to \code{NULL}.}

\item{metricUsed}{A string defining the metric to be used. It can be either \code{mean} or \code{median}. Defaults to \code{median}.}

\item{datasetFolder}{A string defining in which folder to save the results. It can be either \code{full} for the full downsampled dataset, \code{training} for the training downsampled subdataset or \code{validation} for the validation downsampled subdataset. The value must match the origin of the data used. Defaults to \code{full}.}
}
\value{
Generated rds file is saved to \verb{output > 7_Clustering} directory.
}
\description{
This function allows to collapse the phenotypically close cell clusters in order to reduce their number and ease data interpretation. Results are saved in a new \code{rds} file.
}
