% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_Analysis.R
\name{constructPlots}
\alias{constructPlots}
\title{Construct abundance plots for desired features}
\usage{
constructPlots(data = NULL, columnsToPlot = NULL, features = NULL)
}
\arguments{
\item{data}{A data frame containing the dataset to use. This argument should be the output of the \code{bindData()} function. Defaults to \code{NULL}.}

\item{columnsToPlot}{A character vector defining the columns to be used as plotted elements (typically cell clusters). Defaults to \code{NULL}.}

\item{features}{A character vector defining the columns to be used as features (typically group, cluster or metadata) Defaults to \code{NULL}.}
}
\value{
Generated PDF files are saved to \verb{output > 8_Analysis > columnsToPlot} subdirectories.
}
\description{
This function allows to generate and export plots showing the cluster abundance for one or several desired feature(s) of interest. A feature can be for instance a group or a metadata element.
}
