% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Preprocessing.R
\name{convertToRDS}
\alias{convertToRDS}
\title{Convert \code{fcs} files to \code{rds} files}
\usage{
convertToRDS(conversionTable = NULL)
}
\arguments{
\item{conversionTable}{A tabular-delimited text file containing the appropriate information to convert one or several channels to other ones. This is used when the overall staining mix presents the same specificities but slightly different fluorophores depending on the samples. This typically occurs when the cytometer configuration is modified or when the cytometer is different from one batch to another. Defaults to \code{NULL}.

The \code{conversionTable} table follows a pre-defined format: 4 columns in any order (\code{from_desc}, \code{to_desc}, \code{from_name} and \code{to_name}) and any given number of line, each line referring to a specific matching to be treated. For instance, if a line has the values \verb{from_desc = CXCR5 B610-ECD-A}, \verb{to_desc = CXCR5 ECD-A}, \code{from_name = FL2-A} and \code{to_name = FL11-A}, it means that any occurrence of a parameter (in any \code{rds} file) named \code{FL2-A} which also matches the description \verb{CXCR5 B610-ECD-A} will see its values respectively replaced with \code{FL11-A} and \verb{CXCR5 ECD-A}.

Please note that the renaming both affects the parameters of each file AND the parameters in each self-contained compensation matrix.}
}
\value{
Generated \code{rds} files are saved to \code{rds} directory. The function also returns a matrix of all unique parameters (once correctly renamed) used in the dataset.
}
\description{
This function converts every \code{fcs} file in the \code{input} directory to a \code{flowFrame} object (from \code{flowCore} package) encapsulated into a \code{rds} file. It eventually renames the parameters used in the dataset if desired. This step helps to decrease the overall computing time and complexity of several next steps.
}
