% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Transformation.R
\name{transformData}
\alias{transformData}
\title{Transform \code{rds} files}
\usage{
transformData(parametersToTransform = NULL, transformationMethod = "logicle")
}
\arguments{
\item{parametersToTransform}{A vector of all parameters to transform in each \code{rds} file. Defaults to \code{NULL}.}

\item{transformationMethod}{A string defining the transformation method to apply. It can be either \code{logicle} or \code{biexp}. Defaults to \code{logicle}.

Please note that the \code{logicle} transformation needs a file generated using the \code{manualLogicle} R Shiny application which contains the parameters of the \code{logicle} transformation to be used for each parameter. The \code{biexp} transformation does not need this file and can be directly applied.}
}
\value{
Generated \code{rds} files are saved to \code{rds} directory, overwriting the previous ones.
}
\description{
This function applies a mathematical transformation to data from each \code{rds} file.
}
